requirements.txt
openai-whisper==20250625
PyQt6==6.10.2
torch==2.10.0
